requirements.txt
aiohttp==3.7.4.post0
alembic==1.7.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.1
discord.py==1.7.3
fastapi==0.68.1
greenlet==1.1.2
h11==0.12.0
idna==3.2
idna-ssl==1.1.0
importlib-metadata==4.6.4
importlib-resources==5.2.2
Mako==1.1.5
MarkupSafe==2.0.1
multidict==5.1.0
pydantic==1.8.2
python-dotenv==0.19.0
SQLAlchemy==1.4.25
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
yarl==1.6.3
zipp==3.5.0
